Move Pandas columnfile into columnfile.py
