+++ setup.py
@@ -180,7 +180,7 @@ more = [
     "lmfit",       # in sandbox
     "sympy",       # for maths
     'ipywidgets ; python_version >= "3"',  # for notebook nbGui
-    'pyopencl',    # (was?) in sandbox
+    'pyopencl; python_version >= "3"',    # (was?) in sandbox
     'numexpr <= 2.7.0; python_version < "3" ',
     'pyFAI ; python_version >= "3" ',   # pypi problematic
     # 'pyFAI <= 0.18.0 ; python_version  < "3" ',
